requirements.txt
streamlit
python-dotenv
openai
langchain-core
langchain-experimental
langchain-community
base64
streamlit_lottie